Alerts, Animations, Gestures, and Persist Redux Store
When the user submits the form, there is a pop up  confirmation alert. Added another gesture support to Dishdetail component.
When the user fills the reservation form and submits the form, then you will pop up a confirmation Alert with two options: Cancel and OK
Add support for a left to right gesture in the Dishdetail component. When the user swipes from left to right, then you should open the modal containing the comment form.